Started working on PoET

+++ backend/src/main/kotlin/bugs/decentralized/model/Block.kt
@@ -9,7 +9,7 @@ data class Block(
     val timestamp: Long,
     val transactions: List<Transaction>,
     val parentHash: String,
-    val nonce: Int = 0, // a hash that, when combined with the mixHash, proves that the block has gone through proof-of-work
+    val nonce: ULong = 0UL, // proves that the node has waited the necessary amount of time to create a new block
 ) {
 
     private var _hash: String? = null
@@ -20,4 +20,5 @@ data class Block(
 
     private fun computeHash() =
         SHA.sha256(blockNumber.toString() + timestamp + transactions.joinToString("") { it.hash } + parentHash + nonce)
+
 }
